move trial details sub components
