+++ webui/react/src/pages/TrialDetails.tsx
@@ -10,11 +10,11 @@ import Message from 'components/Message';
 import Page from 'components/Page';
 import Section from 'components/Section';
 import Spinner from 'components/Spinner';
-import TrialActions, { Action as TrialAction } from 'components/TrialActions';
-import TrialInfoBox from 'components/TrialInfoBox';
 import handleError, { ErrorType } from 'ErrorHandler';
 import usePolling from 'hooks/usePolling';
 import useRestApi from 'hooks/useRestApi';
+import TrialActions, { Action as TrialAction } from 'pages/TrialDetails/TrialActions';
+import TrialInfoBox from 'pages/TrialDetails/TrialInfoBox';
 import { getExperimentDetails, getTrialDetails, isNotFound } from 'services/api';
 import { TrialDetailsParams } from 'services/types';
 import { ExperimentDetails, TrialDetails } from 'types';
