+++ webui/react/src/pages/TrialDetails/TrialInfoBox.module.scss
